source code
The goal of this project LeoBase is to implement a (very) rudimentary database engine that is loosely based on MySQL.   Through the project, I use visit pattern, factory pattern, bridge pattern, singleton pattern, aslo GRASP to design this project.  This project is implemented by java. It could execute several SQL commands as following:
• SHOW SCHEMAS – Displays all schemas defined in your database.
• USE – Chooses a schema.
• SHOW TABLES – Displays all tables in the currently chosen schema.
• CREATE SCHEMA – Creates a new schema to hold tables.
• CREATE TABLE – Creates a new table schema, i.e. a new empty table.
• INSERT INTO TABLE – Inserts a row/record into a table.
• DROP TABLE – Remove a table schema, and all of its contained data.
• “SELECT-FROM-WHERE” -style query
• EXIT – Cleanly exits the program and saves all table and index information in non-volatile files.